+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.0 - 01/03/2019
+
+- Suggestions box now closes on keyboard hide by default
+- Added property `hideSuggestionsOnKeyboardHide'
+- Width now properly resizes on orientation changes
+- Suggestions box will display above keyboard when keyboard hides the box for AxisDirection.Up
+- Fix FocusNode errors
+- Fix keyboard height calculation
+
 ## 1.0.4/5 - 21/02/2019
 
 - Fix suggestions being called on TextBox focus